requirements.txt
fastapi
uvicorn
boto3
psycopg2-binary
python-dotenv
